seguridad al borrar hdd y lista-hdds

+++ scripts/borrar-hdd.sh
@@ -10,6 +10,14 @@ if [ -z "$INFO" ]; then
 	echo "-1 no se encontro el disco $NAME, exiting"
 	exit
 fi
+BORRABLE=$(grep "^$NAME$"  /home/sperez/maqvirt/lista-hdds.txt)
+if [ -z "$BORRABLE" ]; then
+	echo "-1 no es borrable!"
+	exit
+else
+	S=$(lvremove -f fie_vg/$NAME)
+	# eliminar de la lista y notificar
+	sed -i "/^$NAME$/d" /home/sperez/maqvirt/lista-hdds.txt
+	echo "$S"
+fi
 
-S=$(lvremove -f fie_vg/$NAME)
-echo "$S"
